requirements.txt
flask
python-dotenv
openai
beautifulsoup4